+++ examples/data-sources/meraki_appliance_radio_settings/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_radio_settings" "example" {
+  serial = "1234-ABCD-1234"
+}