+++ examples/resources/meraki_appliance_radio_settings/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_radio_settings.example "<serial>"