+++ docs/guides/changelog.md
@@ -22,6 +22,7 @@ description: |-
 - Add `meraki_appliance_vlans_settings` resource and data source
 - Add `meraki_appliance_prefix_delegated_static` resource and data source
 - Add `meraki_appliance_prefix_delegated_statics` data source
+- Add `meraki_appliance_radio_settings` resource and data source
 
 ## 0.1.1